% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enir.R
\name{enir.build}
\alias{enir.build}
\title{Create ENIR model}
\usage{
enir.build(y, z, scoreFunc = "BIC", alpha = 0.005)
}
\arguments{
\item{y:}{uncalibrated classification scores y [0,1], z: corresponding true class z \{0,1\}}
}
\description{
Create ENIR model
}
\author{
Mahdi Pakdaman
}

