% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enir.R
\name{enir.getMeasures}
\alias{enir.getMeasures}
\title{Computer evaluation measures for binary classification scores}
\usage{
enir.getMeasures(PTE, YTE)
}
\arguments{
\item{PTE:}{vector predicted of classification scores, YTE: corresponding true class \{0,1\}}
}
\description{
Computer evaluation measures for binary classification scores
}
\author{
Mahdi Pakdaman
}

