% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enir.R
\name{enir.predict}
\alias{enir.predict}
\title{predict calibrated estimate for an uncalibrated score using previously built ENIR model}
\usage{
enir.predict(enir, pin, option = 1)
}
\arguments{
\item{y:}{uncalibrated classification scores y [0,1], z: corresponding true class z \{0,1\}}
}
\description{
predict calibrated estimate for an uncalibrated score using previously built ENIR model
}
\author{
Mahdi Pakdaman
}

